Stuck with date conversion. For some reason, the function calendarToDateFormat in Resource will perform a off-by-one error when converting the data.
The problem is the localization.
When looking at what comes out of the DB, Java thinks that it isn't localized, because it assumes that the timezone is different within the DB. Because of this, it tries to jump one month forward, for some really weird reason, instead of just copying the data that was entered.
I don't know how to fix this exactly, because I do not know on which timezone JVM works. I'll try to investigate this.